+++ src/main/java/com/kcc/groo/group/service/GroupService.java
@@ -1,7 +1,10 @@
 package com.kcc.groo.group.service;
 
+import com.kcc.groo.common.exception.GrooException;
 import com.kcc.groo.group.dao.IGroupRepository;
+import com.kcc.groo.group.data.dto.GroupRequestDTO;
 import com.kcc.groo.group.data.model.Groups;
+import com.kcc.groo.group.exception.GroupErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -22,24 +25,41 @@ public class GroupService implements IGroupService {
     /**
      * 새로운 독서 모임 게시글 생성
      *
-     * @param group 독서 모임 게시글 정보
-     * @return 생성된 독서 모임 게시글 정보
      * @author YunSung
      * @created 2025-10-22
+     * @modified 2025-10-23
+     * 게시글 생성 성공/실패 확인 과정 추가
      */
     @Override
-    public Groups createGroup(Groups group) {
+    public Groups createGroup(GroupRequestDTO group, String userId) {
+        // 작성할 독서 모임 게시글 정보 설정
+        Groups newGroup = new Groups();
+        newGroup.setUserId(userId);
+        newGroup.setGroupName(group.getGroupName());
+        newGroup.setBookTitle(group.getBookTitle());
+        newGroup.setHeadcountMin(group.getHeadcountMin());
+        newGroup.setHeadcountMax(group.getHeadcountMax());
+        newGroup.setContent(group.getContent());
+        newGroup.setStyle(group.getStyle());
+        newGroup.setStatus(group.getStatus());
+        newGroup.setEndDate(group.getEndDate());
+        newGroup.setCodeId(group.getCodeId());
+
         // 독서 모임 게시글 생성
-        groupRepository.insertGroup(group);
+        int result = groupRepository.insertGroup(newGroup);
+
+        // 생성 실패 시 예외 발생
+        if (result == 0) {
+            throw new GrooException(GroupErrorCode.FAILED_CREATE_GROUP);
+        }
 
         // 생성된 독서 모임 게시글 정보 반환 (ID 포함)
-        return group;
+        return newGroup;
     }
 
     /**
      * 모든 독서 모임 게시글 목록 조회
      *
-     * @return 독서 모임 게시글 리스트
      * @author YunSung
      * @created 2025-10-22
      */
@@ -52,43 +72,99 @@ public class GroupService implements IGroupService {
     /**
      * 독서 모임 ID를 통한 독서 모임 게시글 상세 조회
      *
-     * @param groupId 조회할 독서 모임 ID
-     * @return 독서 모임 게시글 정보
      * @author YunSung
      * @created 2025-10-22
+     * @modified 2025-10-23
+     * 존재하지 않는 게시글 조회 시 예외 처리 추가
      */
     @Override
     public Groups readGroupByGroupId(int groupId) {
-        // 독서 모임 게시글 조회 및 반환
-        return groupRepository.selectGroupByGroupId(groupId);
+        // 독서 모임 게시글 조회
+        Groups readGroup = groupRepository.selectGroupByGroupId(groupId);
+
+        // 독서 모임 게시글이 존재하지 않을 경우 예외 발생
+        if (readGroup == null) {
+            throw new GrooException(GroupErrorCode.NOT_FOUND_GROUP_TO_READ);
+        }
+
+        // 조회한 독서 모임 게시글 정보 반환
+        return readGroup;
     }
 
     /**
      * 독서 모임 게시글 수정
      *
-     * @param group 수정할 독서 모임 게시글 정보
-     * @return 수정된 독서 모임 게시글 정보
      * @author YunSung
      * @created 2025-10-22
+     * @modified 2025-10-23
+     * 수정할 게시글 존재 여부, 수정 권한, 수정 성공/실패 확인 과정 추가
      */
-    public Groups updateGroup(Groups group) {
+    public Groups updateGroup(GroupRequestDTO group, int groupId, String userId) {
+        // 수정할 독서 모임 게시글 조회
+        Groups targetGroup = groupRepository.selectGroupByGroupId(groupId);
+
+        // 수정할 독서 모임 게시글 ID의 게시글이 존재하지 않으면 예외 발생
+        if (targetGroup == null) {
+            throw new GrooException(GroupErrorCode.NOT_FOUND_GROUP_TO_UPDATE);
+        }
+
+        // 수정할 독서 모임 작성자와 기존 작성자가 다르면 예외 발생
+        if (!targetGroup.getUserId().equals(userId)) {
+            throw new GrooException(GroupErrorCode.FORBIDDEN_UPDATE_GROUP);
+        }
+
+        // 수정할 독서 모임 게시글 정보 설정
+        targetGroup.setGroupName(group.getGroupName());
+        targetGroup.setBookTitle(group.getBookTitle());
+        targetGroup.setHeadcountMin(group.getHeadcountMin());
+        targetGroup.setHeadcountMax(group.getHeadcountMax());
+        targetGroup.setContent(group.getContent());
+        targetGroup.setStyle(group.getStyle());
+        targetGroup.setStatus(group.getStatus());
+        targetGroup.setEndDate(group.getEndDate());
+        targetGroup.setCodeId(group.getCodeId());
+
         // 독서 모임 게시글 수정
-        groupRepository.updateGroup(group);
+        int result = groupRepository.updateGroup(targetGroup);
+
+        // 수정 실패 시 예외 발생
+        if (result == 0) {
+            throw new GrooException(GroupErrorCode.FAILED_UPDATE_GROUP);
+        }
 
         // 수정된 독서 모임 게시글 정보 반환
-        return group;
+        return targetGroup;
     }
 
     /**
      * 특정 독서 모임 게시글 삭제
      *
-     * @param groupId 삭제할 독서 모임 ID
      * @author YunSung
      * @created 2025-10-22
+     * @modified 2025-10-23
+     * 삭제할 게시글 존재 여부, 삭제 권한, 삭제 성공/실패 확인 과정 추가
      */
     @Override
-    public void deleteGroupByGroupId(int groupId) {
+    public void deleteGroupByGroupId(int groupId, String userId) {
+        // 삭제할 독서 모임 게시글 조회
+        Groups deletedGroup = groupRepository.selectGroupByGroupId(groupId);
+
+        // 해당 ID의 독서 모임 게시글이 존재하지 않으면 예외 발생
+        if (deletedGroup == null) {
+            throw new GrooException(GroupErrorCode.NOT_FOUND_GROUP_TO_DELETE);
+        }
+
+        // 해당 ID의 독서 모임 게시글 삭제 작성자와 기존 작성자가 다르면 예외 발생
+        if (!deletedGroup.getUserId().equals(userId)) {
+            throw new GrooException(GroupErrorCode.FORBIDDEN_DELETE_GROUP);
+        }
+
         // 독서 모임 게시글 삭제
-        groupRepository.deleteGroupByGroupId(groupId);
+        int result = groupRepository.deleteGroupByGroupId(groupId);
+
+        // 삭제 실패 시 예외 발생
+        if (result == 0) {
+            throw new GrooException(GroupErrorCode.FAILED_DELETE_GROUP);
+        }
     }
 }
